+++ crates/air_r_formatter/tests/specs/r/roxygen/examples_multiple.R
@@ -1,0 +1,12 @@
+#| [format]
+#| roxygen-examples = true
+
+#' Title
+#'
+#' @param x A value
+#' @examples 1+1
+#' @examples
+#' foo(a,b)
+foo <- function(x) {
+  x
+}