+++ crates/air_r_formatter/tests/specs/r/roxygen/examples_prefix_from_first_line.R
@@ -1,0 +1,9 @@
+#| [format]
+#| roxygen-examples = true
+
+##' Title
+#' @examples
+#' f(a,b)
+foo <- function() {
+  1
+}