+++ crates/air_r_formatter/tests/specs/r/roxygen/examples_tag_line_with_blank_line.R
@@ -1,0 +1,22 @@
+#| [format]
+#| roxygen-examples = true
+
+#' @examples 1 + 1
+#'
+#' 2 + 2
+foo <- function() {
+  1
+}
+
+#' @examples 1 + 1
+#'
+bar <- function() {
+  1
+}
+
+#' @examples 1 + 1
+#'
+#' @returns Something
+baz <- function() {
+  1
+}