+++ crates/air_r_formatter/tests/specs/r/roxygen/examples_trailing_newline.R
@@ -1,0 +1,47 @@
+#| [format]
+#| roxygen-examples = true
+
+#' @examples
+#' 1+1
+foo <- function() {
+  1
+}
+
+#' @examples
+#' 1+1
+#' @returns
+foo <- function() {
+  1
+}
+
+#' @examples
+#' 1+1
+#' 
+foo <- function() {
+  1
+}
+
+#' @examples
+#' 1+1
+#'
+#' @returns 
+foo <- function() {
+  1
+}
+
+#' @examples
+#' 1+1
+#' 
+#'
+foo <- function() {
+  1
+}
+
+#' @examples
+#' 1+1
+#'
+#'
+#' @returns 
+foo <- function() {
+  1
+}